renamed feature constants.py to feature types.py
