+++ tabular/src/autogluon/tabular/data/utils.py
@@ -3,8 +3,8 @@ import math
 
 from pandas import DataFrame
 
-from autogluon.core.features.types import get_type_map_raw
-from autogluon.core.features.constants import R_INT, R_FLOAT, R_CATEGORY
+from autogluon.core.features.infer_types import get_type_map_raw
+from autogluon.core.features.types import R_INT, R_FLOAT, R_CATEGORY
 
 logger = logging.getLogger(__name__)
