+++ tabular/tests/unittests/features/generators/test_identity.py
@@ -1,6 +1,6 @@
 
 from autogluon.tabular.features import IdentityFeatureGenerator
-from autogluon.core.features.constants import R_INT, R_FLOAT
+from autogluon.core.features.types import R_INT, R_FLOAT
 
 
 def test_identity_feature_generator(generator_helper, data_helper):
